+++ info/src/main/java/com/networknt/info/ServerInfoConfig.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.networknt.audit;
+package com.networknt.info;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;