+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Sanitizer middleware to handle cross site scripting on header and body of request.
+- Health endpoint to indicate if the server is alive by sending OK.
 
 ### Changed
 - Update body middleware so that it checks content type and parse the application/json to list or map.
 - Update Validator middleware to handle body as object or string
+- Move info to package info instead of audit
 
 ## 1.0.0 - 2016-10-21